feat: Simplify
Since the code is relatively simple for this app at this stage I moved
the code to the project root and made it executable. So now you execute
using:

   ./ref-cubes cube1 step 1 2.397 0.595

Using Makefile to run and activate didn't work well. To have `make run`
work you needed to do pass the parameters as a environment variable
which you can do most easily by using `=` to assign the parameter to
an environment variable so this works:

  make run parms="cube1 step 1 2.397 0.595"

And then in Makefile this would work:

  run:
    ./rerf-cubes $(parms)

But IMO it's easier to just type:

   ./ref-cubes cube1 step 1 2.397 0.595

As for `make activate` this is not that much harder:

    conda activate cq