feat: Export stl as ascii
With inaccurate advice from ChatGPT I did eventually get it to output
ascii format, although I wonder if an assembly is necessary. For the
moment it works.

Actually I the reason I wanted it a ascii was to see what the vertexes
where as with the binary the 4 boxes are off the plate It seems (0, 0, 0)
isn't in one of the corners of the display at least from Anycubic Workshop
PoV, it seems to think its at centered on one of the short edges. But
Lychee does think it's OK. So we'll see when I get it over to my mac
and try the workshop.