feat: Add boxes-at-corners.pm4m created by Lychee
Note: Anycubic Workshop imported but it was off the build plate
Lychee placed it where I expected. Each block looks to be exactly
in the corner.

Using lychee I added supports to the boxes as without the supports it
wouldn't print when I tried to tweak the printer settings. With the
supports and the base will be 5 layers with 30sec exposures, the
defaults.