feat: Use Context where appropriate
There could be other places but this is the obvious for now and actually
almost all parameters to the routines could be just ctx but that reduces
the generality, I think what we've got so far is a decent start.

Actually, the biggest change was was using ctx.position_box_size and
we are not assuming the notion of a build_plate anywhere.

Currently not using position_box_location, but added a TODO.