feat: Create a prototype BD display solenoid internals
This has 2 cubes on top and bottom with a hollow tube connecting them.
This will allow wire winding's to be wrapped around the column so
when energized a force can be generated to raise and lower the steel
pin that will be attached to the dot and creating a solenoid.