wip: Workaround in support_pillar_upper_cube failing
Apparently The problem is that sweeps with multisection=True can be
problematic and sensitive to small changes.In my case with 3 sections,
bottom, middle and top if I tried to moved the top and bottom locations
towards the origin so the pillars were completely buried the
`wp = wp.sweep(spline_path, multisection=True)` would fail with:
```
OCP.Standard.Standard_NoSuchObject: NCollection_IndexedDataMap::FindFromKey
```

I had a conversation with the bot,
  https://chatgpt.com/share/683338fa-11b0-800c-be81-3cfb4a8a8fb5

And it found that this is a common problem and solutions could be to
move sections, add more sections, remove sections or use `loft` rather
than sweep. After sleeping on it I came up with the idea of removing
the middle section. And that seems to have resolved the problem.